requirements.txt
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
lxml==4.9.2
PyJWT==2.7.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
sqlparse==0.4.4
urllib3==1.26.15
requests==2.28.1
psycopg2-binary==2.9.7